requirements.txt
Flask
Flask-Pymongo
requests
pyquery
